+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Added:
 
 - Support for Wireguard (LinkType RAW) interfaces #29
+- Add str2pcap for improved debugging of logs
 
 ## v0.0.4 - 02-10-2020